requirements.txt
Django==5.2.4
djangorestframework==3.16.0
google-generativeai==0.8.5
python-decouple==3.8
psycopg2-binary==2.9.10
redis==6.2.0
celery==5.5.3
requests==2.32.3
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
msal==1.24.0
websockets==11.0.3
aiortc==1.6.0
aiohttp==3.9.1
channels==4.0.0
channels-redis==4.1.0